+++ src/readRaw.cxx
@@ -1,51 +1,106 @@
-#include <stdio.h>
-
 #include <Common/DataBlock.h>
 #include <Common/DataBlockContainer.h>
 #include <Common/DataSet.h>
 
 #include "RdhUtils.h"
 
+//#define ERRLOG(args...) fprintf(stderr,args)
 
+#include <stdio.h>
+#include <string>
 
-
-typedef struct {
-  uint32_t w0zero : 32;
-  uint32_t w1zero : 32;
-  uint16_t w2zero : 12;
-  uint16_t headerSize : 8;
-  uint16_t linkId : 8;
-  uint16_t feeId : 16;
-  uint16_t blockLength : 16;
-  uint16_t headerVersion : 4;
-} RDHa;
-
-
+#define ERRLOG(args...) fprintf(stderr, args)
 
 int main(int argc, const char *argv[]) {
 
+  // parameters
+  std::string filePath;
+  typedef enum {plain, lz4, undefined} FileType;  
+  FileType fileType=FileType::plain;
+  
+  bool dumpRDH=false;
+  bool validateRDH=true;
+  bool dumpDataBlockHeader=false;
+  int dumpData=0; // if set, number of bytes to dump in each data page
+  
   // parse input arguments
+  // format is a list of key=value pairs
+  
   if (argc<2) {
-    printf("Usage: %s rawFilePath\n",argv[0]);
+    ERRLOG("Usage: %s [rawFilePath] [options]\nList of options:\n \
+    filePath=(string) : path to file\n \
+    dumpRDH=0|1 : dump the RDH headers\n \
+    validateRDH=0|1 : check the RDH headers\n \
+    dumpDataBlockHeader=0|1 : dump the data block headers (internal readout headers)\n \
+    dumpData=(int) : dump the data pages. If -1, all bytes. Otherwise, the first bytes only, as specified.\n",argv[0]);
     return -1;
   }
-  const char *filePath=argv[1];
+   
+  // extra options
+  for (int i=1;i<argc;i++) {
+    const char *option=argv[i];
+    std::string key(option);
+    size_t separatorPosition=key.find('=');
+    if (separatorPosition==std::string::npos) {    
+      // if this is the first argument, use it as the file name (for backward compatibility)
+      if (i==1) {
+        filePath=argv[i];
+      } else { 
+        ERRLOG("Failed to parse option '%s'\n",option);
+      }
+      continue;
+    }
+    key.resize(separatorPosition);
+    std::string value=&(option[separatorPosition+1]);
+    
+    if (key=="fileType") {
+      if (value=="plain") {
+        fileType=FileType::plain;
+      } else if (value=="lz4") {
+        fileType=FileType::lz4;
+      } else {
+        ERRLOG("wrong file type %s\n",value.c_str());
+      }
+    } else if (key == "filePath") {
+      filePath=value;
+    } else if (key == "dumpRDH") {
+      dumpRDH=std::stoi(value);
+    } else if (key == "validateRDH") {
+      validateRDH=std::stoi(value);
+    } else if (key == "dumpDataBlockHeader") {
+      dumpDataBlockHeader=std::stoi(value);
+    } else if (key == "dumpData") {
+      dumpData=std::stoi(value);
+    } else {
+      ERRLOG("unknown option %s\n",key.c_str());
+    }
+    
+  }
+  
+  if (filePath=="") {
+    ERRLOG("Please provide a file name\n");
+    return -1;
+  }
 
+  ERRLOG("Using data file %s\n",filePath.c_str());
+  ERRLOG("dumpRDH=%d validateRDH=%d dumpDataBlockHeader=%d dumpData=%d\n",(int)dumpRDH,(int)validateRDH,(int)dumpDataBlockHeader,dumpData);
+    
   // open raw data file
-  FILE *fp=fopen(filePath,"rb");
+  FILE *fp=fopen(filePath.c_str(),"rb");
   if (fp==NULL) {
-    printf("Failed to open %s\n",filePath);
+    ERRLOG("Failed to open file\n");
     return -1;
   }
-  printf("Reading %s\n",filePath);
   
   // read file
   unsigned long pageCount=0;
-  for(long fileOffset=0;;) {
+  unsigned long RDHBlockCount=0;
+  unsigned long fileOffset=0;
+  for(fileOffset=0;;) {
   
-    #define ERR_LOOP {printf("Error %d @ 0x%08lX\n",__LINE__,fileOffset); break;}
+    #define ERR_LOOP {ERRLOG("Error %d @ 0x%08lX\n",__LINE__,fileOffset); break;}
  
-    long blockOffset=fileOffset;
+    unsigned long blockOffset=fileOffset;
     
     DataBlockHeaderBase hb;
     if (fread(&hb,sizeof(hb),1,fp)!=1) {break;}
@@ -54,49 +109,70 @@ int main(int argc, const char *argv[]) {
     if (hb.blockType!=DataBlockType::H_BASE) {ERR_LOOP;}
     if (hb.headerSize!=sizeof(hb)) {ERR_LOOP;}
     
+    if (dumpDataBlockHeader) {
+      printf("Block header %d @ %lu\n",pageCount+1,fileOffset-sizeof(hb));
+      printf("\tblockType = 0x%02X\n",hb.blockType);
+      printf("\theaderSize = %u\n",hb.headerSize);
+      printf("\tdataSize = %u\n",hb.dataSize);
+      printf("\tid = %lu\n",hb.id);
+      printf("\tlinkId = %u\n",hb.linkId);
+      printf("\tequipmentId = %d\n",(int)hb.equipmentId);
+      printf("\tdata @ %lu\n",fileOffset);
+    }
+    
     void *data=malloc(hb.dataSize);
     if (data==NULL) {ERR_LOOP;}
     if (fread(data,hb.dataSize,1,fp)!=1) {ERR_LOOP;}
     fileOffset+=hb.dataSize;
-    
-    /*
-    printf("Block %ld @ 0x%08lX\n",hb.id,blockOffset);
-    
-    RDH *wPtr=(RDH *)data;
-    for (int i=0;i<RDH_WORDS;i++) {
-      printf("w[%d] = 0x%08X\n", i, wPtr->words[i]);
-    }
-    
-    printf("RDH = %ld bytes\n",sizeof(RDHa));
-    RDHa *hPtr=(RDHa *)data;
-    
-    printf("headerSize=%d\n",(int)hPtr->headerSize);
-    printf("linkId=%d\n",(int)hPtr->linkId);
-    printf("feeId=%d\n",(int)hPtr->feeId);
-    printf("blockLength=%d\n",(int)hPtr->blockLength);
-    printf("headerVersion=%d\n",(int)hPtr->headerVersion);
-    printf("linkId=%d\n",(int)((wPtr->words[2]>>4 ) & 0xFF));
-    */
+    pageCount++;
 
-    std::string errorDescription;    
-    for (size_t pageOffset=0;pageOffset<hb.dataSize;) {
-      pageCount++;
-      RdhHandle h(((uint8_t *)data)+pageOffset);
-      if (h.validateRdh(errorDescription)) {
-        h.dumpRdh();
-        printf("File offset 0x%08lX + %ld\n%s",blockOffset,pageOffset,errorDescription.c_str());
-        errorDescription.clear();
+    if (dumpData) {
+      int max=hb.dataSize;
+      if ((dumpData<max)&&(dumpData>0)) {
+        max=dumpData;
       }
-      pageOffset+=(h.getBlockLength())*(size_t)32; // length counted in 256b words
+      printf("Data page %d @ %lu",pageCount,fileOffset-hb.dataSize);
+      for (int i=0;i<max;i++) {
+        if (i%16==0) {
+          printf("\n\t");
+        }
+        printf("%02X ",(int)(((unsigned char *)data)[i]));
+      }
+      printf ("\n\t...\n");
     }
+
     
+    if ((validateRDH)||(dumpRDH)) {
+      std::string errorDescription;    
+      for (size_t pageOffset=0;pageOffset<hb.dataSize;) {
+        RDHBlockCount++;
+        RdhHandle h(((uint8_t *)data)+pageOffset);
+        
+        if (dumpRDH) {
+          h.dumpRdh();
+        }
+        int nErr=h.validateRdh(errorDescription);
+        if (nErr) {
+          h.dumpRdh();
+          ERRLOG("File offset 0x%08lX + %ld\n%s",blockOffset,pageOffset,errorDescription.c_str());
+          errorDescription.clear();
+          break; // we can not continue decoding if RDH wrong, we are not sure if we can jump to the next ... or should we used fixed 8kB ?
+        }
+        pageOffset+=(h.getBlockLength())*(size_t)32; // length counted in 256b words
+      }
+    }
+        
     free(data);
   }
-  printf("%ld data pages\n",pageCount);
+  ERRLOG("%lu data pages\n",pageCount);
+  if (RDHBlockCount) {
+  ERRLOG("%lu RDH blocks\n",RDHBlockCount);
+  }
+  ERRLOG("%lu bytes\n",fileOffset);
   
   // check file status
   if (feof(fp)) {
-    printf("End of file\n");
+    ERRLOG("End of file\n");
   }
   
   // close file